--- domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/reviewproperty/service/ReviewColumnCreator.java
@@ -1,18 +1,0 @@
-package com.romanticpipe.reviewcanvas.reviewproperty.service;
-
-import org.springframework.stereotype.Service;
-
-import com.romanticpipe.reviewcanvas.reviewproperty.domain.ReviewColumn;
-import com.romanticpipe.reviewcanvas.reviewproperty.repository.ReviewColumnRepository;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class ReviewColumnCreator {
-	private final ReviewColumnRepository reviewColumnRepository;
-
-	public ReviewColumn save(ReviewColumn reviewColumn) {
-		return reviewColumnRepository.save(reviewColumn);
-	}
-}
