+++ domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/reviewproperty/service/ReviewTitleService.java
@@ -16,6 +16,10 @@ public class ReviewTitleService {
 
 	private final ReviewTitleRepository reviewTitleRepository;
 
+	public ReviewTitle save(ReviewTitle reviewTitle) {
+		return reviewTitleRepository.save(reviewTitle);
+	}
+
 	public ReviewTitle validateTitleByShopAdminId(Integer shopAdminId) {
 		return reviewTitleRepository.findByShopAdminIdAndReviewTitleType(shopAdminId, ReviewTitleType.TITLE)
 			.orElseThrow(ReviewTitleNotFoundException::new);
